Create event listener for each button
Create 4 variables to store the 2 numbers, the operator and the solution.
Create 3 variables and assign to an array of buttons with class 'numbers',
array of buttons with class 'operator' and one to update the display
section of the HTML.
Add 'click' event listener for each array. Store the 2 numbers to
operate on in their respective variables. Perform the operation by
calling the operate function. Display the solution each time an
operator button is clicked.