Add percent and +/- features to calculator
Add condition to mutiply the inputted number by 0.01 when user clicks on
the percent button. Similarly add condition to multiply the inputted
number by -1 when user clicks on +/- button.

Add the two conditions within the addDigit function. The function will
return an empty string if the resulting number is equal to 0. Avoid
displaying a number starting with 0 if the user clicks on the percent
or +/- buttons before clicking on a digit.